ml-service/requirements.txt
Flask==3.0.3
Flask-CORS==5.0.0
python-dotenv==1.0.1
psycopg2-binary==2.9.9
scikit-learn==1.5.2
numpy==2.1.3
pandas==2.2.3
requests==2.31.0
ollama==0.1.7
